\name{semPlate-package}
\alias{semPlate-package}
\alias{semPlate}
\docType{package}
\title{
\packageTitle{semPlate}
}
\description{
\packageDescription{semPlate}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{semPlate}
\packageIndices{semPlate}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{semPlate}

Maintainer: \packageMaintainer{semPlate}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
